Tmux based service launch
This adds a tmux and tmuxp based launcher for the backend services
in the docker container. This allows for easy viewing of logs from
the services, as well as attaching and detaching from the logs in
the terminal.

After starting containers with docker-compose up -d, you can
run the following to attach:

docker exec -it karsa_backend_1 \
  /bin/bash -c 'tmux attach -t karsa-services'