Fix docker regression, add persistance and lock file
This fixes a regression in the dockerized setup which results in
failed builds of the electron frontend. This resulted from issues
in electron-updater package failing to fetch a dependency - namely
fs/promises. This is fixed by patching the package, see
karsa-electron/Dockerfile for details.

Additionally, a yarn.lock file was comitted to conform with
standard development practices and help mitigate regressions
in the future.

Finally, this adds persistance to the karsa-backend docker
container by binding volumes for the Projects and TofDaq folders.