Reorganize CSS and do additional cleanup
This attempts to reorganize the CSS for improved developer ergonomics.
CSS is grouped logically; when applicable to a single component, it is
moved to a scoped style section directly in the component. CSS which is
suspected to be deprecated is moved to a new `Deprecated.css` file, but
this file is currently still imported to allow careful evaluation of its
status.

Some vue files are cleaned up a bit and autoformatted.