Move logger from api framework to the store
The logger plugin is current loaded from the API gateway mixin.
This moves the plugin to be loaded directly into the store, since
the logger doesn't concern the API framework in particular and has
broader applicability to the store. This serves to ease logging
configuration, which is expected to be more frequent than changes
to the API framework.