Restructure target and sample stats aggregation
Currently sample and target stats getters have a parameter which
determines whether aggregations are grouped by rating or not. This
removes these parameters in favor of a performing all aggregations
and joining the results together. This simplifies the getter API
and makes the mental model for the stats tables simpler to understand.