Parametrize table height
This adds a table height parameter to BaseTable. It was unclear
how to make the table responsive, and in the first use of this
component in ThePageDataManagement, the correct height was set
by using calc and subtracting the tab height from the viewport
height. Since this depends on the callsite, parametrizing the
height would at least allow the component to be reused even
if it isn't fully responsive.