Fix sample and target data joins and rendering
Some inconsistent statistics have plagued the batch overview
and data management pages due to incorrect joins in the sample
and target stats getters. This fixes the joins by leveraging
the `on x.a = y.b` syntax in place of the `using c` syntax.

Additionally, other minor fixes are applied to ensure correct rendering
of data; columns are reordered to prioritized probable match counts,
and missing values are coalsced to 0.