Add linking columns to data mgmt page
Currently there is no way to link target ions and isotopes to the
compounds they relate to; similarly, matches on ion and isotope
level cannot be linked to the compound level matches they produce.

This adds human-readable compound and ion columns to the target
and match ion and isotope tables in the data management page in
order to provide users with context. Since the same columns are
used for spreadsheet export as the page display, the exported
spreadsheet will also show these columns.