Use expanded buttons in modals
Small buttons are currently used in modals; these float left by default.
Intuitively, confirmation actions should be executed in by a button in
the bottom right corner of a modal. This expands the buttons to make
this the case, and - as a positive sideeffect - make buttons larger and
easier to click.