Add async call to the frontend API
The absence of a way to await asynchronous calls to the backend
prevents us from implementing important performance improvements,
critically it prevents us from being able to sequentially request
match data from the backend easily.

This implemets an asynchronous call - already implemented in the
backend - in the frontend as well.

Related to #2

For usage, see:
  https://gitlab.com/karsa_dev/karsa_msview/-/snippets/2283251