Add context menu to BaseBrowser
The sample and target browser require a variety of actions to be
implemented. This adds a context menu to the BaseBrowser's header
in order to allow such actions to be included.

The menu is added as a prop. Since the menu is positioned in the
header, this deprecates the slot allowing a button to be added in
this place. Previous buttons to trigger target import and batch
creation are moved into this context menu.

Resolves #27