Remove peak data from the frontend
Currently peak data is retrieved to the frontend for all sample items
selected; subsequently, this is sent back to the backend in order to
request matches. Since most of this peak data is not needed in the
frontend otherwise, this results in a significantly slowed performance
without any added benefits.

This removes this unnecessary intermediate step from the frontend in
favor of directly returning matches to the frontend; peaks are instead
retrieved in the backend and never sent to the frontend.