Fix pandas index related bug in matching
Failing to reset the index of pandas dataframes in the matching
functions results in missing matches and exceptions for some data
sets. This fixes the bug by reseting the index of the dataframe.

Note that this reset is performed twice; while this redundency is not
strictly necessary to fix the bug, it is performed defensively in the
event future refactors would reintroduce it.