Directly request matches from the backend
Currently, peak data for all selected samples is retrieved from the
backend to the frontend; however, it is only used to request matches.
This unnecessary data fetching results in a serious performance issue,
making the UI unusuable for standard workflows.

This removes universal peak data loading from the frontend; instead it
unifies the peak and match request endpoints, allowing the frontend to
fetch matches by simply sending a list of target compounds and sample
identifiers.

Resolves #2