Restructure backend and use Poetry for Python depedencies
This simplies the organization of the Python code base; all agents that
need to be installed on dedicated devices are now in their own folders
in the `agents` directory, while the rest of the backend is seperated in
the `backend` folder. Nested packaging is deprecated in favor of a
one-distribution-per-deployment model.

Modern best practices are used for managing the Python package;
Poetry is adopted for efficient and consistent Python dependency
management. pyproject.toml files are used in favor of setup.py.