Transition from Electron to Web Frontend
This redeploys the frontend as a web application instead of as an
electron app. In the process, the bundler is transitioned from Webpack 4
to Vite. To accommidate this, all CJS style require statements are
removed in favor of ESM imports. Additionally, issues with the
vue-plotly library are resolved by porting its source code directly into
the application (it is MIT licensed).