Ensure unmatched rows display no match score tag
Changes to the match API required adjustments to the rendering of
the match score tag in the sample and target browsers. The logic
for whether or not to display the match score tag also needed
updating, and there are still unresolved issues relating to this.

This ensures that match scores are not displayed if the
match_score field is:
- missing
- null
- undefined
- NaN