Change app loading state conditions
Currently the app's loading state - which determins whether to show
a loading spinner - is enabled if no workspace or instrument is focused.
This causes unneccesary loading spinner screens, for example when
switching workspaces. Worse still, it breaks our frontend tests,
since they often try to detect the test workspace is correctly
deleted.

This changes the logic of the loading state so that the application
is loading if and only if there are no workspaces in the store. This
is in line with an intention to transition away from loading screens
and towards skeletons in the UI.