Fixes to the notification store module
The hook provided for cleaning up event listener is erroneously used
in some store modules. Additionally, it was not very clear to other
developers that this hook exists, and that it is used.

This addresses these issues by:
1. Making the hook more explicit by returning it as a method for
  chaining, i.e.
    app.ui.notification
      .on('foo',() => {...})
      .unmount()
2. Ensuring this method is NOT called in stores, but only in
   components; this meant providing a new API for the useMzFit
   composable as well, since it has such a handler and the hook
   is called both in components and stores.

Additionally, the legacy socket event handling API is deprecated
in the store module in favor of directly calling the socket API.
Also, a minor style fix for the notification drawer improves
message spacing.