Unfocus samples before deleting them
Some errors occur when trying to delete samples, because the sample
isn't correctly unfocused before deletion and dependencies on the
focused sample therefore fail to retrieve the right data.

This adds an API to the data module, allowing developers to specify
methods for which we should unfocus before executing. This option is
then used by the sample module for the delete method.

This ensures that unfocusing occurs for all deletion actions across
the frontend. It also allows us to easily unfocus other modules before
calling certain methods.