Improve refocus reliability
Several factors impact the refocusing logic, which is triggered when
reloading a data module and other situations. This attempts to
remedy issues encountered, e.g. when deleting a sample, that
result in focus persisting dispite a record no longer being loaded.

The module's reloading logic's refocusing has been modified to try and
ensure completely unfocusing selecting before reloading. Refocusing
after load is also modified.

Here be dragons: this is probably not the end of the story, but
hopefully for now the major issues are resolved.