API. Fix issue with polarity Validation error
To handle the case where polarity might be None and avoid the ValidationError, modify the SampleFileUpdate model to set a default value of an empty string for polarity. This will ensure that when polarity is None, it will default to an empty string, making it a valid string type.

Debug the calibration. Issue #406