Portable runtimes (#447)
* Allow updating only the CLI

This updates the new windows dev setup powershell script to
allow updating only the CLI or development environment. This
allows a quicker update time for when only part of the code
has changed.

* Implement portable runtime environments

This implements a new system for organizing all the application's
runtime into distinct folders called 'runtimes' which hold all
the state and configuration needed to run an instance of Mascope.

This allows developers to easily switch between different deployments
and configurations. Importantly, it will facilitate managing future
testing and QA environments.

* Remove sphinx build output

* Fix database ops scripts configuration

* Fix gitingnore of runtime folder

This ensures the right files are gitignores from the new runtime
folder:
  - state.json is ignored
  - all runtime folders are ignore
  - only the dev runtimes config is *not* ignored
This ensures developers can share the development environment config
but that all other runtimes are not comitted to the repository.

* Add missing dependencies to Dockerfile

* Exclude legacy and runtime folders from vscode search

* Ensure dev runtime folder structure is committed

* Fix CLI runtime docstrings