Ensure log level is resolve in configuration (#521)
* Ensure log level is resolve in configuration

Currently, while log level is correctly resolved for our logger,
it is not resolved in the configuration. This means that trying
to access runtime.meta.log_level and runtime.config.log_level
can result in errors in many cases.

This moves resolution of log levels from the logger modules of the
runtime to the config module. It ensures log level is always available
in the configuration.

Resolves #522

* Clarify configuration typing

This clarifies Mascope's runtime config typing by renaming the
base class and removing an unnecessary generic from the runtime.

* Document and typehint runtime config code

* Improve autogenerated dev/prod config files