Browser Quick Copy (#527)
* Allow copying fields in sample/target browsers

Using the normal selection and context menu to copy fields
from the sample and target browser is impossible, due to the
custom context menu enabling various functionality.

This adds a new button to copy the contents of the a field
to clipboard. A new component is added to facilitate reusability.

Resolves #517

* Move batch options menu back to table

During the previous commit, I learned how to use the Javascript
event stopPropagation method to prevent buttons in DataTable
from triggering selection.

This allows us to have the batch options menu back on the batch
row, which improves UX by placing it in the right context.