Dev host flag (#532)
* Remove deprecated opts in dev cmd

* Securely expose dev server with --host flag

Previous iterations of the development and production server
have changed the approach for exposing the server over the
network. The latest iteration has actually resulted in
the Vite dev server always being exposed to the network, while
the API was not exposed at all, resulting in a broken app
on other machines.

This adds an explicit CLI option: `mascope dev run --host`,
or -h for short. When this option is set, both the backend
and frontend dev server are exposed to the network. Otherwise,
neither are exposed to the network.

New logging is added to the uvicorn initialization, to make it
clear on what address we are serving. A warning is also emitted
when the `--host` flag is enabled, to ensure developers don't
accidently expose the server.