Batch import ux (#531)
* Frontend: rename sample batch & item to batch & sample

* Batch import: improved validation flows

This improves parsing and error messages for the 'sample type' and
'filter id' fields in the batch import dialog.

For 'sample type', the parser now replaces whitespace with
underscore and capitalizes the string: 'instrument background' is
parsed to 'INSTRUMENT_BACKGROUND'. Additionally, the valid types
are printed in the error message, if the user enters the wrong
type.

For 'filter id' the parser capitalizes the input. The error
messages explains the valid format, and provides examples of
valid ids.

Resolves #529