Batch overview: filter by ionization mechanism (#538)
* Batch overview: filter by ionization mechanism

This enables filtering the batch overview chart by ionization
mechanism, and showing the ion-level match aggregates under that
filter.

  - Filter is applied in the target browser, where a button is now
    available next to the ion-level's mechanism field.
  - Filter is visible as a chip above the batch overview chart,
    and can be removed using the chip's 'remove' button.
  - Filter is automatically removed when switching to a batch
    which doesn't contain that mechanism.

A new store is added to store the filter's state. The match ion
controller is modified to accomidate filtering by ionization
mechanism as well as returning the target ion formula, which is
used for rendering the legend of the chart when the filter is
applied.

Resolves #280

* API: Refactor match_category filtering,add new data to get_match_ions

- Renamed `match_category` to `match_category_min` in query parameters across match-related endpoints to clarify its use in filtering by a specified category and higher (e.g., 1 includes categories 1 and higher).
- Add to `et_match_ions` endpoint `show_ionization_mechanism` and `show_target_ion` flags to allow optional inclusion of related data from the `IonizationMechanism` and `TargetIon` models.
- Updated the frontend `ChartBatchOverview/data.js` to send `match_category_min` for filtering.

Ionization mechanism based filtering in batch overview #280
Missing data in API responses for Jupyter notebooks #493

* API:  Modify get_match_ions for optional compound details

- Addedthe `show_target_compound` flag to the `get_match_ions` endpoint to optionally include related target compound details in the response.

Ionization mechanism based filtering in batch overview #280

* UI: Update match in tooltips and legends.

- Updated hoverbox to include `target_compound_name`, `target_ion_formula`, and conditionally display `ionization_mechanism` for ion level matches. Modified match naming to show compound and ion information in tooltips and legends.

Ionization mechanism based filtering in batch overview #280

* UI: Make ionization mechanism filter button toggleable

Ionization mechanism based filtering in batch overview #280

* UI:  Labels updates

Ionization mechanism based filtering in batch overview #280

* API: Add deduplication flag to get_match_ions and get_match_compounds

- Introduced `deduplicate` parameter to both `get_match_ions` and `get_match_compounds` endpoints to control compound and ion deduplication when `show_target_collection` is set to True, defaults to False.
- Implemented deduplication logic using `deduplicate_match_df` utility function to remove potential duplicates.

Add target focusing in ChartBatchOverview on trace click #542

* UI:  Update trace data

- Modified `queryParams` to retrieve necessary data from the API.
- Added `matchData` object with target IDs (`target_collection_id`, `target_compound_id`, `target_ion_id`) to each trace.
- TODO: Complete `onClick` implementation to correctly focus on match compounds/ions, including handling duplicates.

Add target focusing in ChartBatchOverview on trace click #542

* Enable focusing by predicate function

This allows us to focus matches correctly in the batch
overview chart.

Resolves #542

* Frontend match store key generation & scroll to

The lack of unique identifiers for match compounds, ions and
isotopes produced a variety of bugs relating to focusing them.
This introduces a unique key - generated in the match stores'
load functions - that resolves the issue. The key is then used
in the batch overview chart and target browser to resolve the
bugs.

Additionally, a new functionality is introduces which scrolls
the element to view when focused, improving UX.

* Make target browser active ion mech filter visible

---------

Co-authored-by: Philip Chernonog <philip.chernonog@karsa.fi>