Acquisition tab - file upload, datetime range UX (#526)
* Add file upload functionality to acquisition tab

This adds a file upload button as well as a file dropzone to the
acquisition tab, as well as a new endpoint to facilitate upload.
The user can select on or more h5 or raw files using the system
dialog or drag them into the acqusitions table. Basic error
handeling is implemented, but partial failure modes are not
yet gracefully handled.

A dependency `python-multipart` is added; this is an optional
dependency of FastAPI, necessary to enable form data handling,
which in turn is required for the file upload endpoint.

In the frontend, a minor refactor is also performed to expose
the API's computed URL in the frontend's runtime.

Resolves #397

* Seperate start/end datetime select in acquisition

The PrimeVue datetime range selector has proven to be annoying
in practice. This seperates start and end date/time selection
to enable seperately selecting each. It also fixes the date
format to align with the ISO format used elsewhere in the app.

Resolves #518

* API: Refactor sample_file_upload to controller/route

- Split `sample_file_upload` logic into a controller and route, using `api_decorators`.
- Added file extension validation directly in the route, as using Pydantic models with `UploadFile` is not straightforward.
- Configuration parameters moved to config variables, like allowed file extensions and chunk size.
- A configurable `FILE_UPLOAD_CHUNK_SIZE` (default 2 MB) seems like a reasonable default number, can be adjusted dynamically based on file size if needed.

Implement file upload in the UI #397

* UI: Refactor upload file in modules, add progress

- Module separation: Refactored file upload logic from `sample.js` to `client.js` and `http.js` to improve modularity.
- Progress tracking: Implemented progress notifications for batch file uploads using Axios's `onUploadProgress`, associating `process_id` and `parent_id` to track individual and grouped uploads.
- Extended upload timeout: Increased the file upload timeout (`FILE_UPLOAD_TIMEOUT`) to support larger files and slow network connections.
- Configuration: Moved timeout settings to variables: `API_RESPONSE_TIMEOUT` and `FILE_UPLOAD_TIMEOUT`.

Implement file upload in the UI #397

* API: Add file upload validation Pydantic model

- Implemented file validation for upload size and extension using a Pydantic model (`SampleFileUpload`).
- Utilized FastAPI's `Depends` for validation within the `sample_file_upload_route`.
- Introduced a configurable file size limit (`FILE_UPLOAD_SIZE_LIMIT`) 200 MB and allowed extensions (`FILE_UPLOAD_EXTENSIONS`).
- Enhanced error handling to provide user-friendly error messages for invalid file uploads.

Implement file upload in the UI #397
Improve request validation for Pydantic models #543

* UI: Handle file upload size validation and summary notifications.

- Improved error handling in `upload` method in `frontend/src/api/client.js`:
Added warning and error notifications to summarize the number of successful and failed uploads.
- Added file size validation before upload in `frontend/src/lib/panes/PaneTabAcquisitions.vue`:
Created `validateUploadFiles` function to notify user of oversized files with a single warning.
Configured `FILE_UPLOAD_SIZE_LIMIT` to 200 MB and handled file size validation before proceeding to upload.

Implement file upload in the UI #397

---------

Co-authored-by: Philip Chernonog <philip.chernonog@karsa.fi>