Add --reload option to `mascope dev run` (#535)
This adds a --reload option the `mascope dev run` command in order
to enable HMR for the backend. Due to limitations in Windows an Python,
this is not possible to enable within a single command. See the links
below for details.

As a workaround, the --reload option launches the backend seperately
in a new terminal tab. The user must launch the command from Windows
Terminal for this option to work.

Resolves #498

 https://github.com/encode/uvicorn/discussions/2292
 https://discuss.python.org/t/terminateprocess-via-os-kill-on-windows/30882/5
 https://github.com/metaist/ds/issues/92