Configurable calibration ionization mechanisms (#540)
* Configurable calibration ionization mechanisms

This enables users to seperately configure *calibration* ionization
mechanisms from *matching* ionization mechanisms. Configuration
is performed in the ionization mechanism tab of the edit batch
dialog. A new column is added with a toggle switch for enabling
the use of the mechanism for calibration.

If none are selected, the calibration controller falls back to the
matching mechanisms. The calibration mechanisms must also be
matching mechanisms, and this is enforced by having removing a
mechanism from the calibration mechanism list if it is removed
from the matching list, and by forbidding the user from adding
the mechanism to the calibration list if it is not also added to
the matching list.

Resolves #536

* Fix min isotope abundance param mz fit