Feature: batch overview collection filter (#564)
* Fix version string alignment

* Add collection filters to batch overview chart

This allows to apply multiple collection filters to the batch
overview chart. Filter buttons are added to the target browser,
similar to those used for ionization mechanism filtering previously.
A new array is added to the `app.ui.filter` store to track its
filter state. Chips visualize the filtering state in the batch overview
chart.

Unlike the mechanism filter, this allows for mutliple collections
to filter the chart.

Resolves #563

* Streamline batch overview chart layout

* Fix compound name sorting in the target browser

* Fix target browser scrolling bug

After introducing the scroll-to-selected target feature, the top
of the target browser was obscured after the first selection event.
This adjusts the CSS of the browsers to ensure the contents are
always spaced from the edges of the splitter, fixing the issue.

* Prevent persistence of sample selection

We use features of PrimeVue's DataTable to persist state to local
storage. However, this inadvertently also saves selection, which
in the case of the sample browser is undesirable: it causes load
times for opening a batch to be too long.

This unfocuses the sample when a the state of the table is restored
from local storage. This allows us to keep using the PrimeVue API
for column order and sorting, but reverts prevents selection from
being restored.