Ensure batch import chronology
File selection in the acquisition tab is ordered in order of selection.
This caused confusion in the batch import dialog, since users expect
a chronological order for samples. This ensures the selection
is ordered chronologically before importing samples.

Additionally, columns are added to clarify which file names and
datetime stamps correspond to which records entered by the user.