Nix development environment
This adds a flake.nix enabling development
of Mascope inside Nix-based environments. The
approach is relatively simple, using the
buildFHSEnv helper to create a sandbox
compatible with our Ubuntu deployment mode.

A simplified setup script is added, since much
of the tooling setup in the Ubuntu script can
be offloaded to Nix.