Ensure dev and prod logs go to seperate files
Currently, no seperation between dev and prod logs exists. This is not
ideal since these seperate contexts should be clearly seperated for
debugging purposes. Worse still, this lack of seperation results in a
bug which causes dev logging to fail with a permission error after prod
is run: the docker daemon owns the production log files, and the user
cannot write to them.

This seperates prod and dev logs into distinct folders. Additionally,
file logging from the CLI is altogether disabled. The CLI wasn't
designed to log to file, and having it log to file causes more issues
than it solves.

Resolves #582