Introduce authentication gateway for the app
This leverages the new login and signup components introduced in recent
commits to introduce gateway interface for authentication before
allowing access to the app. Some refactoring is done to facilitate this:
the dashboard is factored out of the MainRoute, and the loading spinner
is moved from the App to the MainRoute. This allows adding the login
and signup forms to the App component, delegating the loading spinner to
the point when the actual dashboard is opened.

Currently no actual protection is offered by this since routes are not
yet protected: data is actually already loaded in the login screen.

Resolves #576