Add loading spinner before user is identified
To prevent a flash of the login screen while the user is identified by
the /users/me endpoint, this adds a loading spinner to the App
component. In the store, unknown user state is null while after the
endpoint returns, if no user is found it is set to false.