Instrument reassign dialog on file upload (#615)
When a file is uploaded, there is currently no validation that the
prefix of the file is a valid instrument in the database. This is
a problem because the backend relies on the file prefix in order
to assign the file to an instrument.

This adds a validation step to check that sample file prefixes are
valid instrument names. If some are invalid, it opens a dialog to
allow the user the chance to assign these to an instrument. When
submitting the form, invalid files are renamed to the selected
instrument, and valid files are kept as they were. All files
are then immediately uploaded.

Closes #575