Match params refactor (#611)
Currently, parameters for matching are called "filter params" since they
relate to filtering the matches to identify possible and probable
matches. This is a bit confusing without context.Additionally, these
params defaults are universal; but optimal defaults for TOF and Orbitrap
instruments differ in practice.

This renames "filter params" to "match params" across the codebase, with
the exception of the database.It also splits the defaults to two
distinguished parameter sets, one for Tof samples and one for Orbi
samples.

In order to ensure the values in the frontend are always in sync with
the backend, an endpoint is created to pass parameters to the frontend.
This endpoint is a generic /api/params endpoint, the idea being that
eventually other parameters can be added here. The filter params
specifically remain in the match models files.

In the frontend, stores are refactored to fit the new conceptual
scheme, and the match params store is modified to load these
defaults, and choose the correct option based on the instrument.

Closes #584