Refactor Frontend API Client
The frontend API client is challenging to maintain: a monolithic
file - api/http.js - wraps all our REST API endpoint calls with
error handlers. This is needless duplication of the same code pattern
but also results in repetition, has the API from here is wrapped a
second time in the store actions.

This refactors the client to centralize error handling using Axios
interceptors, and opts to use the Axios API directly in the store
actions. The hope is that this transparent and flattened API wrapping
would be easier to maintain and understand.

The repsonse handling, error handling and notification logic from the
old client is carried over. Non-standard method calls are enabled with
so-called handlers, which can be selected by passing an `use`
argument to the call config. Similarly, passing a `type` parameter
can set the title for the notification message shown to the user.

Resolves #635