Instruments from `sample_file` table
Currently - in the absence of a dedicated table - instruments in the
app are infered from the `instrument_function` table. We now wish
to implement features allowing users to add new instrument functions,
which requires us to have a different way to infering instruments
*before* they are added to the `instrument_function` table.

This creates a new controller and endpoint which retrieves the
intrument list from the `sample_file` table. The frontend instrument
store is modified to use the new endpoint.

Additionally, inference of instrument type in the backend is centralized
to a single function in `mascope_lib`. This is then used in the new
controller to provide a `type` field in the instrument record.