Improve file upload instrument validation
Since recent changes move towards defining instruments in the database
via the sample files table rather than the instrument functions table,
its important to allow new instruments to be added when uploading files.

This adds support for inputing instruments not listed in the database
when uploading files through the acquisition tab. When implementing
the feature, another issue popped up: it is currently possible to
assign the wrong instrument type, since no validation of the extension
of the file against the prefix is performed.

To address these issue, this commit refactors the file upload
functionality:
 - The functionality is seperated into a distinct component since
   it is growing more complex
 - Validation is refactored into a single forEach loop to make it
   more tractable
 - Seperate checks for the validity of `h5` and `raw` filetypes is
   performed, in addition to the size limit check
 - The dialog allows the user to choose seperate instruments for each
   file type, where input is restricted to valid instruments for that
   type of file
 - The user may also input a new instrument name, and the name is
   validated to conform to the convention for that instrument type