Load frontend store data on login only
Many stores in the frontend currently load data immediately. Now
that authentication is being implemented, this is problematic:
protected routes will fail before authentication.

This adds a hook to the auth store to allow stores and components
to trigger an API call (or any other logic) as soon as a user is
authenticated. It then uses this hook to ensure the stores only
load data after authentication.