Instrument functions pane refactor & dialog
This addressed several issues with the first implementation of the
instrument function pane:
- Improved validation
- Allow updating the fit of existing method files
- Unified submission of instrument function in the sample process dialog
- Preparing the instrument function pane for reuse in other contexts

Additonally, this adds a dialog to allow adjusting instrument functions
for a file, launchable from the sample browser's sample item context
menu. This reuses the instrument function pane developed in previous
commits.

In the backend, endpoints and controllers are modified to facilitate
the UI changes.

Resolves #628