Handle invalid instrument types gracefully
Recently, the way the intrument list is retrieved has been modified to
leverage the sample file table. Additionally, a resolve_instrument_type
function has been created to infer the instrument type from the name of
the instrument.

The resolve function raises an exception when no valid instrument type
can be infered from the name. This causes issues with legacy data, which
may not have been validated.

This adds the option to the resolve function not to raise an exception
in this case, but return None instead. This is used in the
get_instruments controller and invalid instruments are filtered out
of the list.