Restore sample selection when reloading batch
Commit 1ebdf47 introduced a mechanism to the sample browser which
prevents samples from being refocused by the state restoration
mechanism when a batch is reopened. This causes problems when batch
reload events are triggered, for example when rematching a sample.

This adds a mechanism to make an exception in this logic. When the
callback is trigger within 2500ms of a batch reload event, the sample
is refocused.

Towards #657