Secret management for the JWT secret key
This ensures secure management of the JWT secret key, using Docker
Compose's secrets feature. The secret is stored in the following path:

  ${MASCOPE_PATH}/secrets/jwt_secret_key.txt

In development, a hardcoded default is set directly in the code.

This is intended to tentively establish a pattern for managing other
secrets in the app in the future, but the details of the implementation
may yet evolve as we explore this usecase.

Resolves #646