Add API token management to user sidebar
This adds a UI section to the user sidebar drawer to allow users
to regenerate API tokens. The new token is shown in the UI once,
for security reasons; a warning message to users explains this.

A try-except block is added to the backend route, since it raises
an exception if no tokens exist and their removal is attempted.
In the future, if the endpoints are improved, this could probably
be removed in favor a more elegant solution.

Resolves #568