Add MVP user help feature
This adds an in-app user help feature. The user can toggle a 'help mode'
using a button on the top toolbar or via an `alt+h` keyboard shortcut.
In help mode, pertinent info cards supporting HTML rich text are shown
when the mouse is inside some element or component.

A developer-friendly API is introduced using directives for elements,
and the pass-through feature of PrimeVue for library components.Tooltips
are globally hidden when help mode is active.

In order to facilitate documenting dialogs, sidebars and other overlayed
elements, a layers feature is included that allows only help cards in a
certain context to be active.

The API is thoroughly documented in the README.

Resolves #671