Feature: user management dialog
This adds a new dialog for managing users, launchable from the user
sidebar of users with Admin or Owner role. The dialog enables viewing,
modifying, and removing users.

A new user store is added to facitate the feature and provide a central
data model and API for user management. In the backend, new socket emit
calls are added to facilitate reloading user data. The new users store
and the existing auth store are made to react to these events.

A minor fix of the user update controller is also performed to ensure
the backend doesn't throw an error when attempting to modify a user.
Previously, the controller did not check if the username provided is
the same as the current one, and complained the username exists in the
database.

Towards #691