Allow user to change username and password
This adds the ability to edit the username directly in the user sidebar.
A new BaseEditableField component is added to enable a discreet in-place
editing UX. This is used to edit the username directly in the sidebar.

Meanwhile, password reset is enabled through a traditional dialog, which
is lauchable from the sidebar.

Towards #691