Batch import dialog: consolidate CSV logic
This consolidates the parsing and preprocssing logic for autosampler
and generic CSV files into seperate files. Parse functions are extracted
from @/lib/table.js while preprocess functions are extracted from the
BatchImportDialog component.

This brings all the logic next to the only component that uses it, while
seperating the pipelines for generic and autosampler CSV files to make
it easier to understand the each. It also makes the component itself
more understandable.