Fix `free(): invalid pointer` regression on Linux
In Linux development environments, for an unknown reason, recent
changes have caused the `free(): invalid pointer` error to return.

The error originates from the Tofwerk DLLs under certain conditions
for unknown reasons. It is possible the issue is cause by the DLL
being called twice, since the error can refer to trying to free
a memory address which has already been released.

Either way, it seems that removing a line of code previously added to
fix the error, resolved the issue. Since the line is still required to
make the production environment run, it is conditionally executed
only in production.