Simplify instrument config UX and implementation
Since sometime in the dark days of last autumn, we have toiled to
understand what the hell we are doing with instrument functions and
method files in our application.

We shipped one version after another, but alas, we persisted in utter
confusion, as if mocked by the gods of chaos. We couldn't keep track
of neither the implementation, nor the UX.

Finally we decided enough is enough. With radical simplification as our
sharp blade, we cut free of the tangle and saw the light.

Previously, we allows multiple instrument functions to have the same
method_file name, and allowed the user to update a method file by
refitting new instrument functions.

No more, we said!

Instrument functions and method files are now one. One record to rule
them all, and in the light bind them. The new unified concept is to be
called 'instrument configs'. We shall bow down in awe before it.

Instrument configs are immutable. No more worrying about updates. Each
instrument config within an instrument has a unique method_file (a.k.a
'instrument config name'). No more dread of how method_files are to
map to instrument functions.

By the grace of the gods of clean code, we rejoice!

The instrument functions API is updated to use the 'new' convention
and renamed to instrument configs. The new implementation enforces
the uniqueness constaint in the controllers; eventually, this will
be done in the database. Oh, then we shall be truely enlightened!

In the frontend the UI implementation is significantly simplified.
Naming is also made sane. Users can now look at the dialogs without
going mad and worshiping Cthulu.

Resolves #738

Or maybe it doesn't, and I am trapped in a purgatory where I must
reimplement this feature again and again and again and again....
...forever.